Actualización del código y corregir varios errores
