+++ src/main/java/es/javier/timerPlugin/commands/SetTimerCommand.java
@@ -8,7 +8,13 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
 public class SetTimerCommand implements CommandExecutor {
+
+    private final Map<UUID, TimerConfig> playerSelections = new ConcurrentHashMap<>();
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, @NotNull String @NotNull [] args) {
          if (command.getName().equalsIgnoreCase("settimer")) {
@@ -43,5 +49,23 @@ public class SetTimerCommand implements CommandExecutor {
             }
             return true;
         }
+        return false;
+    }
+
+    private int convertToSeconds(int time, String unit) {
+        switch (unit) {
+            case "s": return time;
+            case "m": return time * 60;
+            case "h": return time * 3600;
+            case "d": return time * 86400;
+            default: return 60;
+        }
+    }
+
+    private String formatTime(int seconds) {
+        if (seconds < 60) return seconds + " segundos";
+        if (seconds < 3600) return (seconds / 60) + " minutos y " + (seconds % 60) + " segundos";
+        if (seconds < 86400) return (seconds / 3600) + " horas, " + ((seconds % 3600) / 60) + " minutos";
+        return (seconds / 86400) + " días, " + ((seconds % 86400) / 3600) + " horas";
     }
 }
